Never overwrite files with directories by default
There is a 'OverwriteDestinationFiles' option in the configuration file.
The purpose of this option was to give guidance in the situations where
you would overwrite files with files, which is a common enough thing to
do. However, I think it's a stretch to use apply this option to
overwriting files with *directories*. For instance let's say a user had
a file on his drive called 'a', and was transfering over directories
that happened to include the name 'a'. Would the user expect his file to
be overwritten in this case, config option or no? My guess is no. It's
almost always a mistake to overwrite files with directories, so I'm not
going to have 'OverwriteDestinationFiles' apply to such a situation.