Fix logic error in finding device and mount locs
Before I was testing if the mount point was a subset of the destination
string. This is not sufficient, because, say if I had mount points z/
and zebrafish/ and that I wanted to transfer to zebrafish/, I might end
up transfering to z/ instead since both are subsets of the destination
string I would provide. Now this is fixed :-)